+++ lib/src/widgets/solid_preferences_models.dart
@@ -187,6 +187,7 @@ class SolidAppBarActionIds {
   static const String about = 'about';
   static const String logout = 'logout';
   static const String preferences = 'preferences';
+  static const String inviteOthers = 'invite_others';
 
   SolidAppBarActionIds._();
 }